requirements.txt
numpy
pandas
scikit-learn
joblib
flask
paho-mqtt
requests
tensorflow==2.17.0 ; python_version >= '3.9' and python_version < '3.13'
